requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
google-generativeai>=0.8.0
pypdf>=4.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
rich>=13.7.0
pandas>=2.2.0
tabulate>=0.9.0